Add `example_web` app
