+++ check.sh
@@ -9,7 +9,10 @@ CARGO_INCREMENTAL=0 cargo clippy --workspace --all-targets --all-features --  -D
 cargo test --workspace --all-targets --all-features
 cargo test --workspace --doc
 
+cargo check -p egui --target wasm32-unknown-unknown
+cargo check -p egui_web --target wasm32-unknown-unknown
 cargo check -p demo_web --target wasm32-unknown-unknown
+cargo check -p example_web --target wasm32-unknown-unknown
 
 # For finding bloat:
 # cargo bloat --release --bin demo_glium -n 200 | rg egui
