+++ demo_glium/src/main.rs
@@ -1,3 +1,4 @@
+#![forbid(unsafe_code)]
 #![deny(warnings)]
 #![warn(clippy::all)]
