+++ docs/example.html
@@ -28,7 +28,10 @@
 </head>
 
 <body>
-    <!-- The WASM code will resize this to cover the entire screen -->
+    <!--
+        THis is where the app will show up.
+        The WASM code will resize this to cover the entire screen.
+    -->
     <canvas id="the_canvas_id"></canvas>
 
     <script>
@@ -45,13 +48,13 @@
     </script>
 
     <!-- this is the JS generated by the `wasm-bindgen` CLI tool -->
-    <script src="demo_web.js"></script>
+    <script src="example_web.js"></script>
 
     <script>
         // We'll defer our execution until the wasm is ready to go.
         // Here we tell bindgen the path to the wasm file so it can start
         // initialization and return to us a promise when it's done.
-        wasm_bindgen("./demo_web_bg.wasm")
+        wasm_bindgen("./example_web_bg.wasm")
             .then(on_wasm_loaded)["catch"](console.error);
 
         function on_wasm_loaded() {
