+++ example_glium/src/main.rs
@@ -1,63 +1,10 @@
 //! Example of how to use Egui
-
+#![forbid(unsafe_code)]
 #![deny(warnings)]
 #![warn(clippy::all)]
 
-/// We derive Deserialize/Serialize so we can persist app state on shutdown.
-#[derive(serde::Deserialize, serde::Serialize)]
-struct MyApp {
-    name: String,
-    age: u32,
-}
-
-impl Default for MyApp {
-    fn default() -> Self {
-        Self {
-            name: "Arthur".to_owned(),
-            age: 42,
-        }
-    }
-}
-
-impl egui::app::App for MyApp {
-    /// Called each time the UI needs repainting, which may be many times per second.
-    /// Put your widgets into a `SidePanel`, `TopPanel`, `CentralPanel`, `Window` or `Area`.
-    fn ui(
-        &mut self,
-        ctx: &std::sync::Arc<egui::Context>,
-        integration_context: &mut egui::app::IntegrationContext,
-    ) {
-        let MyApp { name, age } = self;
-
-        // Example used in `README.md`.
-        egui::CentralPanel::default().show(ctx, |ui| {
-            ui.heading("My Egui Application");
-
-            ui.horizontal(|ui| {
-                ui.label("Your name: ");
-                ui.text_edit_singleline(name);
-            });
-
-            ui.add(egui::Slider::u32(age, 0..=120).text("age"));
-            if ui.button("Click each year").clicked {
-                *age += 1;
-            }
-
-            ui.label(format!("Hello '{}', age {}", name, age));
-
-            ui.advance_cursor(16.0);
-            if ui.button("Quit").clicked {
-                integration_context.output.quit = true;
-            }
-        });
-
-        integration_context.output.window_size = Some(ctx.used_size()); // resize the window to be just the size we need it to be
-    }
-
-    fn on_exit(&mut self, storage: &mut dyn egui::app::Storage) {
-        egui::app::set_value(storage, egui::app::APP_KEY, self);
-    }
-}
+mod example_app;
+use example_app::ExampleApp;
 
 fn main() {
     let title = "My Egui Window";
@@ -68,6 +15,8 @@ fn main() {
     // Alternative: store nowhere
     // let storage = egui::app::DummyStorage::default();
 
-    let app: MyApp = egui::app::get_value(&storage, egui::app::APP_KEY).unwrap_or_default(); // Restore `MyApp` from file, or create new `MyApp`.
+    // Restore `example_app` from file, or create new `ExampleApp`:
+    let app: ExampleApp = egui::app::get_value(&storage, egui::app::APP_KEY).unwrap_or_default();
+
     egui_glium::run(title, Box::new(storage), app);
 }
