+++ example_web/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "example_web"
+version = "0.1.0"
+authors = ["Emil Ernerfeldt <emil.ernerfeldt@gmail.com>"]
+license = "MIT OR Apache-2.0"
+edition = "2018"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+egui = { path = "../egui", features = ["serde"] }
+egui_web = { path = "../egui_web" }
+js-sys = "0.3"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+wasm-bindgen = "0.2"
