+++ example_web/src/lib.rs
@@ -2,6 +2,8 @@
 #![deny(warnings)]
 #![warn(clippy::all)]
 
+mod example_app;
+
 use wasm_bindgen::prelude::*;
 
 /// This is the entry-point for all the web-assembly.
@@ -10,7 +12,7 @@ use wasm_bindgen::prelude::*;
 /// You can add more callbacks like this if you want to call in to your code.
 #[wasm_bindgen]
 pub fn start(canvas_id: &str) -> Result<(), wasm_bindgen::JsValue> {
-    let app = egui::DemoApp::default();
+    let app = example_app::ExampleApp::default();
     let backend = egui_web::WebBackend::new(canvas_id)?;
     let runner = egui_web::AppRunner::new(backend, Box::new(app))?;
     egui_web::start(runner)?;
